+++ packages/natives/native/index.js
@@ -39,7 +39,7 @@ export const VcsRepo = nativeBindings.VcsRepo;
 
 // functions
 export const __ompInstallTokioRuntime = nativeBindings.__ompInstallTokioRuntime;
-export const __piNativesV1_8_2 = nativeBindings.__piNativesV1_8_2;
+export const __piNativesV1_8_3 = nativeBindings.__piNativesV1_8_3;
 export const astEdit = nativeBindings.astEdit;
 export const astGrep = nativeBindings.astGrep;
 export const astMatch = nativeBindings.astMatch;